Fixing stricter compile warnings; Using xccode configs.

+++ Classes/sha1.c
@@ -120,7 +120,7 @@ void SHA1Update(SHA1_CTX* context, unsigned char* data, unsigned int len)
 {
   unsigned int i, j;
   
-  j = (context->count[0] >> 3) & 63;
+  j = (unsigned int)((context->count[0] >> 3) & 63);
   if ((context->count[0] += len << 3) < (len << 3)) context->count[1]++;
   context->count[1] += (len >> 29);
   if ((j + len) > 63) {
